Change float to CGFloat

+++ Rubik's Timer/RubiksUtil.m
@@ -234,9 +234,9 @@
 }
 
 + (UIColor *)randomComplimentColor:(UIColor *)color {
-    float hue, sat, brightness, alpha;
+    CGFloat hue, sat, brightness, alpha;
     [color getHue:&hue saturation:&sat brightness:&brightness alpha:&alpha];
-    float newHue = ((int)(hue * 360) + 270) % 360 / 360.0f;
+    CGFloat newHue = ((int)(hue * 360) + 270) % 360 / 360.0f;
     return [UIColor colorWithHue:newHue saturation:0.8f brightness:1.0f alpha:0.75f];
 }
